Optmized storage management

+++ src/helpers/firebase/file.js
@@ -1,7 +1,6 @@
 const {bucket} = require('../../../config/firebase');
 const uploadFile = async (file,filepath) =>{
-    const blob = bucket.file(filepath);
-
+    const blob = bucket.file(filepath);    
     await new Promise((resolve, reject) => {
       const blobStream = blob.createWriteStream({
         metadata: { contentType: file.mimetype }
@@ -19,6 +18,9 @@ const uploadFile = async (file,filepath) =>{
 const deleteFile = async (prefix) => {
   try {
     const [files] = await bucket.getFiles({ prefix });    
+    if (files.length === 0) {
+      return { success: false, message: 'File not found' };
+    }
     await Promise.all(files.map(file => file.delete()));
   } catch (error) {
     return { success: false, message: error.message };
